requirements.txt
pandas
Pillow
ultralytics
opencv-python
numpy
requests
beautifulsoup4
selenium
webdriver-manager
scikit-image
matplotlib
